set up front-end

+++ backend/controllers/product.controller.js
@@ -46,13 +46,16 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
-    console.log("id:",id);
+    console.log("id:", id);
+
+    if (!mongoose.Types.ObjectId.isValid(id))
+        return res.status(404).json({ success: false, message: "Invalid product id" });
 
     try {
         let _ = await Product.findByIdAndDelete(id);
         res.json(201).json({ success: true, message: `Element with id: ${id} successfully deleted.` })
     } catch (error) {
         console.error('Error occured in delete by id:', error.message);
-        res.status(404).json({ success: false, message: "Product not found." });
+        res.status(500).json({ success: false, message: "Server error" });
     }
 };
